Library updates + support directAccessUrl
- update only to latest Jetty 11 - further updates require significant refactoring
- support newer Alfresco Transform Model, including directAccessUrl option
- enhance transform test form

+++ base/src/main/java/de/acosix/alfresco/transform/base/impl/ContextImpl.java
@@ -150,13 +150,16 @@ public class ContextImpl implements Context
     @Override
     public void discardTempFile(final Path tempFile)
     {
-        try
+        if (tempFile != null)
         {
-            Files.deleteIfExists(tempFile);
-        }
-        catch (final IOException ignore)
-        {
-            // ignore
+            try
+            {
+                Files.deleteIfExists(tempFile);
+            }
+            catch (final IOException ignore)
+            {
+                // ignore
+            }
         }
     }
